+++ datasets/base_det_dataset.py
@@ -1,0 +1,494 @@
+import os
+
+import mmcv
+import numpy as np
+import torch
+from mmdet3d.core.bbox.structures.lidar_box3d import LiDARInstance3DBoxes
+from PIL import Image
+from pyquaternion import Quaternion
+from torch.utils.data import Dataset
+
+from .utils import bev_transform, depth_transform, img_transform
+
+
+class BaseDetDataset(Dataset):
+    def __init__(self,
+                 ida_aug_conf,
+                 bda_aug_conf,
+                 classes,
+                 data_root,
+                 info_paths,
+                 is_train,
+                 use_cbgs=False,
+                 num_sweeps=1,
+                 img_conf=dict(img_mean=[123.675, 116.28, 103.53],
+                               img_std=[58.395, 57.12, 57.375],
+                               to_rgb=True),
+                 return_depth=False,
+                 sweep_idxes=list(),
+                 key_idxes=list(),
+                 use_fusion=False):
+        """Dataset used for bevdetection task.
+        Args:
+            ida_aug_conf (dict): Config for ida augmentation.
+            bda_aug_conf (dict): Config for bda augmentation.
+            classes (list): Class names.
+            use_cbgs (bool): Whether to use cbgs strategy,
+                Default: False.
+            num_sweeps (int): Number of sweeps to be used for each sample.
+                default: 1.
+            img_conf (dict): Config for image.
+            return_depth (bool): Whether to use depth gt.
+                default: False.
+            sweep_idxes (list): List of sweep idxes to be used.
+                default: list().
+            key_idxes (list): List of key idxes to be used.
+                default: list().
+            use_fusion (bool): Whether to use lidar data.
+                default: False.
+        """
+        super().__init__()
+        if isinstance(info_paths, list):
+            self.infos = list()
+            for info_path in info_paths:
+                self.infos.extend(mmcv.load(info_path))
+        else:
+            self.infos = mmcv.load(info_paths)
+        self.is_train = is_train
+        self.ida_aug_conf = ida_aug_conf
+        self.bda_aug_conf = bda_aug_conf
+        self.data_root = data_root
+        self.classes = classes
+        self.use_cbgs = use_cbgs
+        if self.use_cbgs:
+            self.cat2id = {name: i for i, name in enumerate(self.classes)}
+            self.sample_indices = self._get_sample_indices()
+        self.num_sweeps = num_sweeps
+        self.img_mean = np.array(img_conf['img_mean'], np.float32)
+        self.img_std = np.array(img_conf['img_std'], np.float32)
+        self.to_rgb = img_conf['to_rgb']
+        self.return_depth = return_depth
+        assert sum([sweep_idx >= 0 for sweep_idx in sweep_idxes]) \
+            == len(sweep_idxes), 'All `sweep_idxes` must greater \
+                than or equal to 0.'
+
+        self.sweeps_idx = sweep_idxes
+        assert sum([key_idx < 0 for key_idx in key_idxes]) == len(key_idxes),\
+            'All `key_idxes` must less than 0.'
+        self.key_idxes = [0] + key_idxes
+        self.use_fusion = use_fusion
+
+    def _get_sample_indices(self, name_map=None):
+        """Load annotations from ann_file.
+
+        Args:
+            ann_file (str): Path of the annotation file.
+
+        Returns:
+            list[dict]: List of annotations after class sampling.
+        """
+        class_sample_idxs = {cat_id: [] for cat_id in self.cat2id.values()}
+        for idx, info in enumerate(self.infos):
+            gt_names = set(
+                [ann_info['category_name'] for ann_info in info['ann_infos']])
+            for gt_name in gt_names:
+                if name_map is not None:
+                    gt_name = name_map[gt_name]
+                if gt_name not in self.classes:
+                    continue
+                class_sample_idxs[self.cat2id[gt_name]].append(idx)
+        duplicated_samples = sum(
+            [len(v) for _, v in class_sample_idxs.items()])
+        class_distribution = {
+            k: len(v) / duplicated_samples
+            for k, v in class_sample_idxs.items()
+        }
+
+        sample_indices = []
+
+        frac = 1.0 / len(self.classes)
+        ratios = [frac / v for v in class_distribution.values()]
+        for cls_inds, ratio in zip(list(class_sample_idxs.values()), ratios):
+            sample_indices += np.random.choice(cls_inds,
+                                               int(len(cls_inds) *
+                                                   ratio)).tolist()
+        return sample_indices
+
+    def sample_ida_augmentation(self):
+        """Generate ida augmentation values based on ida_config."""
+        H, W = self.ida_aug_conf['H'], self.ida_aug_conf['W']
+        fH, fW = self.ida_aug_conf['final_dim']
+        if self.is_train:
+            resize = np.random.uniform(*self.ida_aug_conf['resize_lim'])
+            resize_dims = (int(W * resize), int(H * resize))
+            newW, newH = resize_dims
+            crop_h = int(
+                (1 - np.random.uniform(*self.ida_aug_conf['bot_pct_lim'])) *
+                newH) - fH
+            crop_w = int(np.random.uniform(0, max(0, newW - fW)))
+            crop = (crop_w, crop_h, crop_w + fW, crop_h + fH)
+            flip = False
+            if self.ida_aug_conf['rand_flip'] and np.random.choice([0, 1]):
+                flip = True
+            rotate_ida = np.random.uniform(*self.ida_aug_conf['rot_lim'])
+        else:
+            resize = max(fH / H, fW / W)
+            resize_dims = (int(W * resize), int(H * resize))
+            newW, newH = resize_dims
+            crop_h = int(
+                (1 - np.mean(self.ida_aug_conf['bot_pct_lim'])) * newH) - fH
+            crop_w = int(max(0, newW - fW) / 2)
+            crop = (crop_w, crop_h, crop_w + fW, crop_h + fH)
+            flip = False
+            rotate_ida = 0
+        return resize, resize_dims, crop, flip, rotate_ida
+
+    def sample_bda_augmentation(self):
+        """Generate bda augmentation values based on bda_config."""
+        if self.is_train:
+            rotate_bda = np.random.uniform(*self.bda_aug_conf['rot_lim'])
+            scale_bda = np.random.uniform(*self.bda_aug_conf['scale_lim'])
+            flip_dx = np.random.uniform() < self.bda_aug_conf['flip_dx_ratio']
+            flip_dy = np.random.uniform() < self.bda_aug_conf['flip_dy_ratio']
+        else:
+            rotate_bda = 0
+            scale_bda = 1.0
+            flip_dx = False
+            flip_dy = False
+        return rotate_bda, scale_bda, flip_dx, flip_dy
+
+    def get_image(self, cam_infos, cams, lidar_infos=None):
+        """Given data and cam_names, return image data needed.
+
+        Args:
+            sweeps_data (list): Raw data used to generate the data we needed.
+            cams (list): Camera names.
+
+        Returns:
+            Tensor: Image data after processing.
+            Tensor: Transformation matrix from camera to ego.
+            Tensor: Intrinsic matrix.
+            Tensor: Transformation matrix for ida.
+            Tensor: Transformation matrix from key
+                frame camera to sweep frame camera.
+            Tensor: timestamps.
+            dict: meta infos needed for evaluation.
+        """
+        assert len(cam_infos) > 0
+        sweep_imgs = list()
+        sweep_sensor2ego_mats = list()
+        sweep_intrin_mats = list()
+        sweep_ida_mats = list()
+        sweep_sensor2sensor_mats = list()
+        sweep_timestamps = list()
+        sweep_lidar_depth = list()
+        if self.return_depth or self.use_fusion:
+            sweep_lidar_points = list()
+            for lidar_info in lidar_infos:
+                lidar_path = lidar_info['LIDAR_TOP']['filename']
+                lidar_points = np.fromfile(os.path.join(
+                    self.data_root, lidar_path),
+                                           dtype=np.float32,
+                                           count=-1).reshape(-1, 5)[..., :4]
+                sweep_lidar_points.append(lidar_points)
+        for cam in cams:
+            imgs = list()
+            sensor2ego_mats = list()
+            intrin_mats = list()
+            ida_mats = list()
+            sensor2sensor_mats = list()
+            timestamps = list()
+            lidar_depth = list()
+            key_info = cam_infos[0]
+            resize, resize_dims, crop, flip, \
+                rotate_ida = self.sample_ida_augmentation(
+                    )
+            for sweep_idx, cam_info in enumerate(cam_infos):
+
+                img = Image.open(
+                    os.path.join(self.data_root, cam_info[cam]['filename']))
+                # img = Image.fromarray(img)
+                w, x, y, z = cam_info[cam]['calibrated_sensor']['rotation']
+                # sweep sensor to sweep ego
+                sweepsensor2sweepego_rot = torch.Tensor(
+                    Quaternion(w, x, y, z).rotation_matrix)
+                sweepsensor2sweepego_tran = torch.Tensor(
+                    cam_info[cam]['calibrated_sensor']['translation'])
+                sweepsensor2sweepego = sweepsensor2sweepego_rot.new_zeros(
+                    (4, 4))
+                sweepsensor2sweepego[3, 3] = 1
+                sweepsensor2sweepego[:3, :3] = sweepsensor2sweepego_rot
+                sweepsensor2sweepego[:3, -1] = sweepsensor2sweepego_tran
+                # sweep ego to global
+                w, x, y, z = cam_info[cam]['ego_pose']['rotation']
+                sweepego2global_rot = torch.Tensor(
+                    Quaternion(w, x, y, z).rotation_matrix)
+                sweepego2global_tran = torch.Tensor(
+                    cam_info[cam]['ego_pose']['translation'])
+                sweepego2global = sweepego2global_rot.new_zeros((4, 4))
+                sweepego2global[3, 3] = 1
+                sweepego2global[:3, :3] = sweepego2global_rot
+                sweepego2global[:3, -1] = sweepego2global_tran
+
+                # global sensor to cur ego
+                w, x, y, z = key_info[cam]['ego_pose']['rotation']
+                keyego2global_rot = torch.Tensor(
+                    Quaternion(w, x, y, z).rotation_matrix)
+                keyego2global_tran = torch.Tensor(
+                    key_info[cam]['ego_pose']['translation'])
+                keyego2global = keyego2global_rot.new_zeros((4, 4))
+                keyego2global[3, 3] = 1
+                keyego2global[:3, :3] = keyego2global_rot
+                keyego2global[:3, -1] = keyego2global_tran
+                global2keyego = keyego2global.inverse()
+
+                # cur ego to sensor
+                w, x, y, z = key_info[cam]['calibrated_sensor']['rotation']
+                keysensor2keyego_rot = torch.Tensor(
+                    Quaternion(w, x, y, z).rotation_matrix)
+                keysensor2keyego_tran = torch.Tensor(
+                    key_info[cam]['calibrated_sensor']['translation'])
+                keysensor2keyego = keysensor2keyego_rot.new_zeros((4, 4))
+                keysensor2keyego[3, 3] = 1
+                keysensor2keyego[:3, :3] = keysensor2keyego_rot
+                keysensor2keyego[:3, -1] = keysensor2keyego_tran
+                keyego2keysensor = keysensor2keyego.inverse()
+                keysensor2sweepsensor = (
+                    keyego2keysensor @ global2keyego @ sweepego2global
+                    @ sweepsensor2sweepego).inverse()
+                sweepsensor2keyego = global2keyego @ sweepego2global @\
+                    sweepsensor2sweepego
+                sensor2ego_mats.append(sweepsensor2keyego)
+                sensor2sensor_mats.append(keysensor2sweepsensor)
+                intrin_mat = torch.zeros((4, 4))
+                intrin_mat[3, 3] = 1
+                intrin_mat[:3, :3] = torch.Tensor(
+                    cam_info[cam]['calibrated_sensor']['camera_intrinsic'])
+                if self.return_depth and (self.use_fusion or sweep_idx == 0):
+                    point_depth = self.get_lidar_depth(
+                        sweep_lidar_points[sweep_idx], img,
+                        lidar_infos[sweep_idx], cam_info[cam])
+                    point_depth_augmented = depth_transform(
+                        point_depth, resize, self.ida_aug_conf['final_dim'],
+                        crop, flip, rotate_ida)
+                    lidar_depth.append(point_depth_augmented)
+                img, ida_mat = img_transform(
+                    img,
+                    resize=resize,
+                    resize_dims=resize_dims,
+                    crop=crop,
+                    flip=flip,
+                    rotate=rotate_ida,
+                )
+                ida_mats.append(ida_mat)
+                img = mmcv.imnormalize(np.array(img), self.img_mean,
+                                       self.img_std, self.to_rgb)
+                img = torch.from_numpy(img).permute(2, 0, 1)
+                imgs.append(img)
+                intrin_mats.append(intrin_mat)
+                timestamps.append(cam_info[cam]['timestamp'])
+            sweep_imgs.append(torch.stack(imgs))
+            sweep_sensor2ego_mats.append(torch.stack(sensor2ego_mats))
+            sweep_intrin_mats.append(torch.stack(intrin_mats))
+            sweep_ida_mats.append(torch.stack(ida_mats))
+            sweep_sensor2sensor_mats.append(torch.stack(sensor2sensor_mats))
+            sweep_timestamps.append(torch.tensor(timestamps))
+            if self.return_depth:
+                sweep_lidar_depth.append(torch.stack(lidar_depth))
+        # Get mean pose of all cams.
+        ego2global_rotation = np.mean(
+            [key_info[cam]['ego_pose']['rotation'] for cam in cams], 0)
+        ego2global_translation = np.mean(
+            [key_info[cam]['ego_pose']['translation'] for cam in cams], 0)
+        img_metas = dict(
+            box_type_3d=LiDARInstance3DBoxes,
+            ego2global_translation=ego2global_translation,
+            ego2global_rotation=ego2global_rotation,
+        )
+
+        ret_list = [
+            torch.stack(sweep_imgs).permute(1, 0, 2, 3, 4),
+            torch.stack(sweep_sensor2ego_mats).permute(1, 0, 2, 3),
+            torch.stack(sweep_intrin_mats).permute(1, 0, 2, 3),
+            torch.stack(sweep_ida_mats).permute(1, 0, 2, 3),
+            torch.stack(sweep_sensor2sensor_mats).permute(1, 0, 2, 3),
+            torch.stack(sweep_timestamps).permute(1, 0),
+            img_metas,
+        ]
+        if self.return_depth:
+            ret_list.append(torch.stack(sweep_lidar_depth).permute(1, 0, 2, 3))
+        return ret_list
+
+    def get_gt(self, info, cams):
+        pass
+
+    def choose_cams(self):
+        """Choose cameras randomly.
+
+        Returns:
+            list: Cameras to be used.
+        """
+        if self.is_train and self.ida_aug_conf['Ncams'] < len(
+                self.ida_aug_conf['cams']):
+            cams = np.random.choice(self.ida_aug_conf['cams'],
+                                    self.ida_aug_conf['Ncams'],
+                                    replace=False)
+        else:
+            cams = self.ida_aug_conf['cams']
+        return cams
+
+    def __getitem__(self, idx):
+        if self.use_cbgs:
+            idx = self.sample_indices[idx]
+        cam_infos = list()
+        lidar_infos = list()
+        # TODO: Check if it still works when number of cameras is reduced.
+        cams = self.choose_cams()
+        for key_idx in self.key_idxes:
+            cur_idx = key_idx + idx
+            # Handle scenarios when current idx doesn't have previous key
+            # frame or previous key frame is from another scene.
+            if cur_idx < 0:
+                cur_idx = idx
+            elif self.infos[cur_idx]['scene_token'] != self.infos[idx][
+                    'scene_token']:
+                cur_idx = idx
+            info = self.infos[cur_idx]
+            cam_infos.append(info['cam_infos'])
+            lidar_infos.append(info['lidar_infos'])
+            lidar_sweep_timestamps = [
+                lidar_sweep['LIDAR_TOP']['timestamp']
+                for lidar_sweep in info['lidar_sweeps']
+            ]
+            for sweep_idx in self.sweeps_idx:
+                if len(info['cam_sweeps']) == 0:
+                    cam_infos.append(info['cam_infos'])
+                    lidar_infos.append(info['lidar_infos'])
+                else:
+                    # Handle scenarios when current sweep doesn't have all
+                    # cam keys.
+                    for i in range(min(len(info['cam_sweeps']) - 1, sweep_idx),
+                                   -1, -1):
+                        if sum([cam in info['cam_sweeps'][i]
+                                for cam in cams]) == len(cams):
+                            cam_infos.append(info['cam_sweeps'][i])
+                            cam_timestamp = np.mean([
+                                val['timestamp']
+                                for val in info['cam_sweeps'][i].values()
+                            ])
+                            # Find the closest lidar frame to the cam frame.
+                            lidar_idx = np.abs(lidar_sweep_timestamps -
+                                               cam_timestamp).argmin()
+                            lidar_infos.append(info['lidar_sweeps'][lidar_idx])
+                            break
+        if self.return_depth or self.use_fusion:
+            image_data_list = self.get_image(cam_infos, cams, lidar_infos)
+
+        else:
+            image_data_list = self.get_image(cam_infos, cams)
+        ret_list = list()
+        (
+            sweep_imgs,
+            sweep_sensor2ego_mats,
+            sweep_intrins,
+            sweep_ida_mats,
+            sweep_sensor2sensor_mats,
+            sweep_timestamps,
+            img_metas,
+        ) = image_data_list[:7]
+        img_metas['token'] = self.infos[idx]['sample_token']
+        if self.is_train:
+            gt_boxes, gt_labels = self.get_gt(self.infos[idx], cams)
+        # Temporary solution for test.
+        else:
+            gt_boxes = sweep_imgs.new_zeros(0, 7)
+            gt_labels = sweep_imgs.new_zeros(0, )
+
+        rotate_bda, scale_bda, flip_dx, flip_dy = self.sample_bda_augmentation(
+        )
+        bda_mat = sweep_imgs.new_zeros(4, 4)
+        bda_mat[3, 3] = 1
+        gt_boxes, bda_rot = bev_transform(gt_boxes, rotate_bda, scale_bda,
+                                          flip_dx, flip_dy)
+        bda_mat[:3, :3] = bda_rot
+        ret_list = [
+            sweep_imgs,
+            sweep_sensor2ego_mats,
+            sweep_intrins,
+            sweep_ida_mats,
+            sweep_sensor2sensor_mats,
+            bda_mat,
+            sweep_timestamps,
+            img_metas,
+            gt_boxes,
+            gt_labels,
+        ]
+        if self.return_depth:
+            ret_list.append(image_data_list[7])
+        return ret_list
+
+    def __str__(self):
+        pass
+
+    def __len__(self):
+        if self.use_cbgs:
+            return len(self.sample_indices)
+        else:
+            return len(self.infos)
+
+
+def collate_fn(data, is_return_depth=False):
+    imgs_batch = list()
+    sensor2ego_mats_batch = list()
+    intrin_mats_batch = list()
+    ida_mats_batch = list()
+    sensor2sensor_mats_batch = list()
+    bda_mat_batch = list()
+    timestamps_batch = list()
+    gt_boxes_batch = list()
+    gt_labels_batch = list()
+    img_metas_batch = list()
+    depth_labels_batch = list()
+    for iter_data in data:
+        (
+            sweep_imgs,
+            sweep_sensor2ego_mats,
+            sweep_intrins,
+            sweep_ida_mats,
+            sweep_sensor2sensor_mats,
+            bda_mat,
+            sweep_timestamps,
+            img_metas,
+            gt_boxes,
+            gt_labels,
+        ) = iter_data[:10]
+        if is_return_depth:
+            gt_depth = iter_data[10]
+            depth_labels_batch.append(gt_depth)
+        imgs_batch.append(sweep_imgs)
+        sensor2ego_mats_batch.append(sweep_sensor2ego_mats)
+        intrin_mats_batch.append(sweep_intrins)
+        ida_mats_batch.append(sweep_ida_mats)
+        sensor2sensor_mats_batch.append(sweep_sensor2sensor_mats)
+        bda_mat_batch.append(bda_mat)
+        timestamps_batch.append(sweep_timestamps)
+        img_metas_batch.append(img_metas)
+        gt_boxes_batch.append(gt_boxes)
+        gt_labels_batch.append(gt_labels)
+    mats_dict = dict()
+    mats_dict['sensor2ego_mats'] = torch.stack(sensor2ego_mats_batch)
+    mats_dict['intrin_mats'] = torch.stack(intrin_mats_batch)
+    mats_dict['ida_mats'] = torch.stack(ida_mats_batch)
+    mats_dict['sensor2sensor_mats'] = torch.stack(sensor2sensor_mats_batch)
+    mats_dict['bda_mat'] = torch.stack(bda_mat_batch)
+    ret_list = [
+        torch.stack(imgs_batch),
+        mats_dict,
+        torch.stack(timestamps_batch),
+        img_metas_batch,
+        gt_boxes_batch,
+        gt_labels_batch,
+    ]
+    if is_return_depth:
+        ret_list.append(torch.stack(depth_labels_batch))
+    return ret_list
