+++ datasets/utils.py
@@ -1,0 +1,117 @@
+import numpy as np
+import torch
+from PIL import Image
+
+
+def get_rot(h):
+    return torch.Tensor([
+        [np.cos(h), np.sin(h)],
+        [-np.sin(h), np.cos(h)],
+    ])
+
+
+def img_transform(img, resize, resize_dims, crop, flip, rotate):
+    ida_rot = torch.eye(2)
+    ida_tran = torch.zeros(2)
+    # adjust image
+    img = img.resize(resize_dims)
+    img = img.crop(crop)
+    if flip:
+        img = img.transpose(method=Image.FLIP_LEFT_RIGHT)
+    img = img.rotate(rotate)
+
+    # post-homography transformation
+    ida_rot *= resize
+    ida_tran -= torch.Tensor(crop[:2])
+    if flip:
+        A = torch.Tensor([[-1, 0], [0, 1]])
+        b = torch.Tensor([crop[2] - crop[0], 0])
+        ida_rot = A.matmul(ida_rot)
+        ida_tran = A.matmul(ida_tran) + b
+    A = get_rot(rotate / 180 * np.pi)
+    b = torch.Tensor([crop[2] - crop[0], crop[3] - crop[1]]) / 2
+    b = A.matmul(-b) + b
+    ida_rot = A.matmul(ida_rot)
+    ida_tran = A.matmul(ida_tran) + b
+    ida_mat = ida_rot.new_zeros(4, 4)
+    ida_mat[3, 3] = 1
+    ida_mat[2, 2] = 1
+    ida_mat[:2, :2] = ida_rot
+    ida_mat[:2, 3] = ida_tran
+    return img, ida_mat
+
+
+def bev_transform(gt_boxes, rotate_angle, scale_ratio, flip_dx, flip_dy):
+    rotate_angle = torch.tensor(rotate_angle / 180 * np.pi)
+    rot_sin = torch.sin(rotate_angle)
+    rot_cos = torch.cos(rotate_angle)
+    rot_mat = torch.Tensor([[rot_cos, -rot_sin, 0], [rot_sin, rot_cos, 0],
+                            [0, 0, 1]])
+    scale_mat = torch.Tensor([[scale_ratio, 0, 0], [0, scale_ratio, 0],
+                              [0, 0, scale_ratio]])
+    flip_mat = torch.Tensor([[1, 0, 0], [0, 1, 0], [0, 0, 1]])
+    if flip_dx:
+        flip_mat = flip_mat @ torch.Tensor([[-1, 0, 0], [0, 1, 0], [0, 0, 1]])
+    if flip_dy:
+        flip_mat = flip_mat @ torch.Tensor([[1, 0, 0], [0, -1, 0], [0, 0, 1]])
+    rot_mat = flip_mat @ (scale_mat @ rot_mat)
+    if gt_boxes.shape[0] > 0:
+        gt_boxes[:, :3] = (rot_mat @ gt_boxes[:, :3].unsqueeze(-1)).squeeze(-1)
+        gt_boxes[:, 3:6] *= scale_ratio
+        gt_boxes[:, 6] += rotate_angle
+        if flip_dx:
+            gt_boxes[:, 6] = 2 * torch.asin(torch.tensor(1.0)) - gt_boxes[:, 6]
+        if flip_dy:
+            gt_boxes[:, 6] = -gt_boxes[:, 6]
+        if gt_boxes.shape[1] > 7:
+            gt_boxes[:, 7:] = (
+                rot_mat[:2, :2] @ gt_boxes[:, 7:].unsqueeze(-1)).squeeze(-1)
+    return gt_boxes, rot_mat
+
+
+def depth_transform(cam_depth, resize, resize_dims, crop, flip, rotate):
+    """Transform depth based on ida augmentation configuration.
+
+    Args:
+        cam_depth (np array): Nx3, 3: x,y,d.
+        resize (float): Resize factor.
+        resize_dims (list): Final dimension.
+        crop (list): x1, y1, x2, y2
+        flip (bool): Whether to flip.
+        rotate (float): Rotation value.
+
+    Returns:
+        np array: [h/down_ratio, w/down_ratio, d]
+    """
+
+    H, W = resize_dims
+    cam_depth[:, :2] = cam_depth[:, :2] * resize
+    cam_depth[:, 0] -= crop[0]
+    cam_depth[:, 1] -= crop[1]
+    if flip:
+        cam_depth[:, 0] = resize_dims[1] - cam_depth[:, 0]
+
+    cam_depth[:, 0] -= W / 2.0
+    cam_depth[:, 1] -= H / 2.0
+
+    h = rotate / 180 * np.pi
+    rot_matrix = [
+        [np.cos(h), np.sin(h)],
+        [-np.sin(h), np.cos(h)],
+    ]
+    cam_depth[:, :2] = np.matmul(rot_matrix, cam_depth[:, :2].T).T
+
+    cam_depth[:, 0] += W / 2.0
+    cam_depth[:, 1] += H / 2.0
+
+    depth_coords = cam_depth[:, :2].astype(np.int16)
+
+    depth_map = np.zeros(resize_dims)
+    valid_mask = ((depth_coords[:, 1] < resize_dims[0])
+                  & (depth_coords[:, 0] < resize_dims[1])
+                  & (depth_coords[:, 1] >= 0)
+                  & (depth_coords[:, 0] >= 0))
+    depth_map[depth_coords[valid_mask, 1],
+              depth_coords[valid_mask, 0]] = cam_depth[valid_mask, 2]
+
+    return torch.Tensor(depth_map)
