+++ exps/nuscenes/base_exp.py
@@ -12,7 +12,8 @@ from pytorch_lightning.core import LightningModule
 from torch.cuda.amp.autocast_mode import autocast
 from torch.optim.lr_scheduler import MultiStepLR
 
-from datasets.nusc_det_dataset import NuscDetDataset, collate_fn
+from datasets.base_det_dataset import collate_fn
+from datasets.nusc_det_dataset import NuscDetDataset
 from evaluators.det_evaluators import DetNuscEvaluator
 from models.base_bev_depth import BaseBEVDepth
 from utils.torch_dist import all_gather_object, get_rank, synchronize
