+++ exps/nuscenes/fusion/bev_depth_fusion_lss_r50_256x704_128x128_24e.py
@@ -5,7 +5,8 @@ import torch.utils.data
 import torch.utils.data.distributed
 
 from exps.base_cli import run_cli
-from exps.base_exp import BEVDepthLightningModel as BaseBEVDepthLightningModel
+from exps.nuscenes.base_exp import \
+    BEVDepthLightningModel as BaseBEVDepthLightningModel
 from models.fusion_bev_depth import FusionBEVDepth
