+++ exps/nuscenes/mv/bev_depth_lss_r50_256x704_128x128_24e_2key.py
@@ -23,7 +23,8 @@ traffic_cone    0.462   0.516   0.355   nan     nan     nan
 barrier 0.512   0.491   0.275   0.200   nan     nan
 """
 from exps.base_cli import run_cli
-from exps.base_exp import BEVDepthLightningModel as BaseBEVDepthLightningModel
+from exps.nuscenes.base_exp import \
+    BEVDepthLightningModel as BaseBEVDepthLightningModel
 from models.base_bev_depth import BaseBEVDepth
