+++ exps/nuscenes/mv/bev_depth_lss_r50_256x704_128x128_24e_2key_ema.py
@@ -22,8 +22,8 @@ traffic_cone    0.453   0.519   0.335   nan     nan     nan
 barrier 0.506   0.475   0.279   0.178   nan     nan
 """
 from exps.base_cli import run_cli
-from exps.mv.bev_depth_lss_r50_256x704_128x128_24e_2key import \
-    BEVDepthLightningModel
+
+from .bev_depth_lss_r50_256x704_128x128_24e_2key import BEVDepthLightningModel
 
 if __name__ == '__main__':
     run_cli(BEVDepthLightningModel,
